requirements.txt
amqp==2.6.1
asgiref==3.2.10
backcall==0.2.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
decorator==4.4.2
Django==3.1.2
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.3.0
django-timezone-field==4.0
django-treebeard==4.3.1
djangorestframework==3.12.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
hachoir==3.1.1
html5lib==1.1
idna==2.10
ImageHash==4.1.0
ipython==7.18.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
kombu==4.6.11
l18n==2018.5
numpy==1.19.2
openpyxl==3.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
requests==2.24.0
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
traitlets==5.0.5
Unidecode==1.1.1
urllib3==1.25.10
vine==1.3.0
wagtail==2.10.2
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
